opcua: adding the new feature as "OPC-UA Industrial Edge Demo".
In this demo, the UAExpert works as OPC UA Client and PolarFire SoC
ICICLE Kit and PolarFire SoC Video Kit acts as OPC UA Servers,
to communicate over OPC protocol.

The communication is structured on a client-server architecture
that defines the interaction between client and server, where the client
initiates the communication and the server provides the requested data.

Signed-off-by: Ahmed Mujtaba Ali <ahmed.ali2@microchip.com>